Theoretically removed the cycle, but still getting an error when deleting an account.

+++ DataLayer/Migrations/Scripts/GameEvent_Up.sql
@@ -56,7 +56,8 @@ AS
   BEGIN
     DELETE FROM GameEventParticipation
 	FROM DELETED
-	WHERE GameEventParticipation.TakesPartInId = DELETED.Id;
+	WHERE GameEventParticipation.TakesPartInId = DELETED.Id
+	  AND GameEventParticipation.ParticipationStatus <> 'Creator';
 
 	DELETE FROM Game -- This DELETE could be replaced with CascadeDelete on Games, but it seems more consistent to put everything in one place.
 	FROM DELETED
